Semi-persistent notifications
Notifications can be sent to players that are both online and offline.
If a player is online, it'll be displayed directly, when he is offline, the notification is saved.
When this player logs in, the number of unread notifications will be displayed and he can access them all by typing /nt or /notification